refactor: Piece.getScore의 반환 타입을 double로 변경

+++ src/main/java/domain/board/Board.java
@@ -46,7 +46,7 @@ public final class Board {
         return alivePieces.hasRoyalPiece(side);
     }
 
-    public int getTotalScore(Side side) {
+    public double getTotalScore(Side side) {
         return alivePieces.getTotalScore(side);
     }
