+++ src/test/java/domain/piece/GeneralTest.java
@@ -56,10 +56,10 @@ class GeneralTest {
     void 본인의_점수를_반환한다() {
         // given
         General general = new General(Side.HAN);
-        int expected = 0;
+        double expected = 0;
 
         // when
-        int actual = general.getScore();
+        double actual = general.getScore();
 
         // then
         assertThat(actual).isEqualTo(expected);
